Feature/Made Add to cart a component and rendered on single product page
